+++ lib/modules/perf/perf_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: perf
-// Updated at: 2026-06-05 21:13:25
-// Commit: test(perf): implement widget integration tests for perf
+// Updated at: 2026-06-05 22:44:51
+// Commit: refactor(perf): extract reusable atomic widgets from perf
 
 class perfComponentState {
-  final String commitRef = '1790';
-  final String lastMessage = 'test(perf): implement widget integration tests for perf';
-  final DateTime timestamp = DateTime.parse('2026-06-05T21:13:25+05:00');
+  final String commitRef = '1796';
+  final String lastMessage = 'refactor(perf): extract reusable atomic widgets from perf';
+  final DateTime timestamp = DateTime.parse('2026-06-05T22:44:51+05:00');
 }